// frontend/src/app/core/api/model/course-session-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CourseSessionResponse { 
    id?: number;
    sessionDate?: string;
    startTime?: string;
    endTime?: string;
    sessionType?: CourseSessionResponse.SessionTypeEnum;
    status?: CourseSessionResponse.StatusEnum;
    lessonTitle?: string;
    courseContent?: string;
    cancellationReason?: string;
    classGroupId?: number;
    className?: string;
    specialty?: string;
    subjectId?: number;
    subjectName?: string;
    teacherId?: number;
    teacherName?: string;
    roomId?: number;
    roomName?: string;
    formattedDate?: string;
    attendanceSubmitted?: boolean;
}
export namespace CourseSessionResponse {
    export const SessionTypeEnum = {
        Regular: 'REGULAR',
        Replacement: 'REPLACEMENT',
        Support: 'SUPPORT',
        ExamSupervision: 'EXAM_SUPERVISION'
    } as const;
    export type SessionTypeEnum = typeof SessionTypeEnum[keyof typeof SessionTypeEnum];
    export const StatusEnum = {
        Planned: 'PLANNED',
        InProgress: 'IN_PROGRESS',
        Completed: 'COMPLETED',
        Canceled: 'CANCELED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}