// frontend/src/app/core/api/model/course-session-search-request.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CourseSessionSearchRequest { 
    startTime?: string;
    endTime?: string;
    promotionId?: number;
    teacherId?: number;
    subjectId?: number;
    status?: CourseSessionSearchRequest.StatusEnum;
    isAttendanceSubmitted?: boolean;
    type?: CourseSessionSearchRequest.TypeEnum;
}
export namespace CourseSessionSearchRequest {
    export const StatusEnum = {
        Planned: 'PLANNED',
        InProgress: 'IN_PROGRESS',
        Completed: 'COMPLETED',
        Canceled: 'CANCELED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
    export const TypeEnum = {
        Regular: 'REGULAR',
        Replacement: 'REPLACEMENT',
        Support: 'SUPPORT',
        ExamSupervision: 'EXAM_SUPERVISION'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}